requirements.txt
Flask
pdfminer.six
python-docx
requests
Pillow
python-dotenv
